+++ wazo_confd/config.py
@@ -1,4 +1,4 @@
-# Copyright 2013-2024 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2013-2025 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import argparse
@@ -128,6 +128,7 @@ DEFAULT_CONFIG = {
         'queue_general': True,
         'queue_member': True,
         'queue_schedule': True,
+        'recordings': True,
         'register_iax': True,
         'registrar': True,
         'rtp': True,